+++ packages/types/src/index.ts
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 export * from "./api/index.js";
+export * from "./client/index.js";
 export * from "./config/index.js";
 export * from "./contract/index.js";
 export * from "./datasets/index.js";
